+++ ganttproject/src/biz/ganttproject/impex/csv/CsvWriter.java
@@ -1,0 +1,15 @@
+package biz.ganttproject.impex.csv;
+
+
+import biz.ganttproject.impex.common.GanttExportWriter;
+
+/**
+ * Created with IntelliJ IDEA.
+ * User: akurutin
+ * Date: 03.04.17
+ * Time: 19:41
+ */
+public interface CsvWriter extends GanttExportWriter {
+
+
+}
