+++ ganttproject/src/biz/ganttproject/impex/xls/XlsWriter.java
@@ -1,0 +1,9 @@
+package biz.ganttproject.impex.xls;
+
+import biz.ganttproject.impex.common.GanttExportWriter;
+
+/**
+ * Created by Маша on 04.04.2017.
+ */
+public interface XlsWriter extends GanttExportWriter {
+}
