+++ ganttproject/src/biz/ganttproject/impex/xls/XlsWriterImpl.java
@@ -1,0 +1,35 @@
+package biz.ganttproject.impex.xls;
+
+import java.io.IOError;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.io.UnsupportedEncodingException;
+
+/**
+ * Created by Маша on 04.04.2017.
+ */
+public class XlsWriterImpl implements XlsWriter {
+    public XlsWriterImpl(OutputStream stream) throws IOException {
+        throw new UnsupportedEncodingException();
+    }
+
+    @Override
+    public void print(String value) throws IOException {
+        throw new UnsupportedEncodingException();
+    }
+
+    @Override
+    public void println() throws IOException {
+        throw new UnsupportedEncodingException();
+    }
+
+    @Override
+    public void flush() throws IOException {
+        throw new UnsupportedEncodingException();
+    }
+
+    @Override
+    public void close() throws IOException {
+        throw new UnsupportedEncodingException();
+    }
+}
